refactor: moved timestamp creation to a separate util module for clarity
